refactored multiplet function into their own py file
