+++ doc/documentation.rst
@@ -38,7 +38,7 @@ Reference manual
    reference/solve_parameters
    reference/moves
    reference/measurements
-   reference/utils
+   reference/multiplet
 
 FAQs
 ----
