+++ doc/reference/multiplet.rst
@@ -1,6 +1,6 @@
 Utility functions
 =================
 
-.. automodule:: triqs_cthyb.util
+.. automodule:: triqs_cthyb.multiplet_tools
    :members:
    :show-inheritance:
